Cast 8 bits variable to allow left shift by 8
Type of csd.v2.c_size_mid is uint8_t (8 bits). In order to left shift by 8 bits csd.v2.c_size_mid without losing informations (bits), is necessary to cast the variable to uint32_t.

+++ src/utility/Sd2Card.cpp
@@ -164,7 +164,7 @@ uint32_t Sd2Card::cardSize(void) {
     return (uint32_t)(c_size + 1) << (c_size_mult + read_bl_len - 7);
   } else if (csd.v2.csd_ver == 1) {
     uint32_t c_size = ((uint32_t)csd.v2.c_size_high << 16)
-                      | (csd.v2.c_size_mid << 8) | csd.v2.c_size_low;
+                      | ((uint32_t)csd.v2.c_size_mid << 8) | csd.v2.c_size_low;
     return (c_size + 1) << 10;
   } else {
     error(SD_CARD_ERROR_BAD_CSD);
